Use GraphQL/Apollo in AssetAdmin (fixes #316)
At the moment this is limited to "read files"
and "create folder". The remaining API endpoints will be
successively moved over to GraphQL.

The GraphQL type creators are likely temporary,
to be replaced with a shorter scaffolding-based version:
https://github.com/silverstripe/silverstripe-graphql/issues/9
https://github.com/silverstripe/silverstripe-graphql/issues/22
https://github.com/silverstripe/silverstripe-graphql/issues/23

RFC at https://github.com/silverstripe/silverstripe-framework/issues/6245

+++ client/src/containers/Editor/Editor.js
@@ -31,7 +31,10 @@ class Editor extends Component {
     }
 
     if (name === 'action_delete') {
-      this.props.onDelete(data.ID);
+      // eslint-disable-next-line no-alert
+      if (confirm(i18n._t('AssetAdmin.CONFIRMDELETE'))) {
+        this.props.onDelete(data.ID);
+      }
       event.preventDefault();
       return;
     }
